Fixed CM lines so they draw correctly. Made them rectangles instead of lines and supplied a custom determineBounds()

+++ ideal-gas-v3/test-src/TestMoleculeBoxCollision.java
@@ -45,7 +45,7 @@ public class TestMoleculeBoxCollision extends PhetApplication {
                                               new Vector2D.Double( 0, 0 ),
                                               new Vector2D.Double(),
                                               5 );
-            m.setVelocity( new Vector2D.Double( 0, 200 ) );
+            m.setVelocity( new Vector2D.Double( 200, 300 ) );
             PumpMoleculeCmd cmd = new PumpMoleculeCmd( (IdealGasModel)getModel(),
                                                        m,
                                                        this );
